requirements.txt
colorama==0.4.6
numpy==1.23.4
Pillow==9.3.0
